+++ src/ca/powerj/NForecast.java
@@ -162,9 +162,9 @@ class NForecast extends NBase {
 	}
 
 	@Override
-	void setFilter(byte[] rows) {
+	void setFilter(int[] rows) {
 		for (int i = 0; i < rowsView.length; i++) {
-			rowsView[i] = rows[i];
+			rowsView[i] = (byte) rows[i];
 		}
 		altered = true;
 	}
